December release of minor bug fixes

+++ swift/70000TonsBands/iCloudFileHandler.swift
@@ -9,6 +9,7 @@
 import Foundation
 
 
+
 func createiCloudDirectory(){
     do {
         if let iCloudDocumentsURL = FileManager.default.url(forUbiquityContainerIdentifier:nil)?.appendingPathComponent("Documents"){
@@ -49,63 +50,69 @@ func copyDocumentsToiCloudDrive() {
 }
 
 func saveFileToiCloudDrive(localFile : URL, fileName : String){
+    
+    if (savingiCloud == false){
+        savingiCloud = true;
 
-    createiCloudDirectory()
-    DispatchQueue.global(qos: DispatchQoS.QoSClass.default).async {
-        
-        let iCloudDocumentsURL = FileManager.default.url(forUbiquityContainerIdentifier:nil)?.appendingPathComponent("Documents").appendingPathComponent(fileName, isDirectory: false)
-        
-        if (iCloudDocumentsURL?.absoluteString.isEmpty == false){
-            do {
-                if (iCloudDocumentsURL?.path != nil){
-                    if (FileManager.default.fileExists(atPath: (iCloudDocumentsURL?.path)!) == true){
-                        //try FileManager.default.removeItem(atPath: iCloudDocumentsURL!.path)
+        createiCloudDirectory()
+        DispatchQueue.global(qos: DispatchQoS.QoSClass.default).async {
+            
+            let iCloudDocumentsURL = FileManager.default.url(forUbiquityContainerIdentifier:nil)?.appendingPathComponent("Documents").appendingPathComponent(fileName, isDirectory: false)
+            
+            if (iCloudDocumentsURL?.absoluteString.isEmpty == false){
+                do {
+                    if (iCloudDocumentsURL?.path != nil){
+                        if (FileManager.default.fileExists(atPath: (iCloudDocumentsURL?.path)!) == true){
+                            try FileManager.default.copyItem(atPath:localFile.path, toPath: iCloudDocumentsURL!.path)
+                            print("iCloud Drive wrote file \(iCloudDocumentsURL?.path)")
+                            try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
+                        }
                     }
-                
-                    print("iCloud Drive saving file from \(localFile) and copying to \(iCloudDocumentsURL?.path)")
-                    try FileManager.default.copyItem(atPath:localFile.path, toPath: iCloudDocumentsURL!.path)
-                    print("iCloud Drive wrote file \(iCloudDocumentsURL?.path)")
-                    try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
+                } catch {
+                    print("iCloud Drive error on load of single file of \(localFile.path) could not be copied to \(iCloudDocumentsURL?.path) " + error.localizedDescription);
                 }
-            } catch {
-                print("iCloud Drive error on load of single file of \(localFile.path) could not be copied to \(iCloudDocumentsURL?.path) " + error.localizedDescription);
+            } else {
+                print("iCloud Drive file \(fileName) is not available in the cloud")
             }
-        } else {
-            print("iCloud Drive file \(fileName) is not available in the cloud")
         }
+        savingiCloud = false;
     }
 }
 
 
 func loadFileFromiCloudDrive(localFile : URL, fileName : String){
 
-    let iCloudDocumentsURL = FileManager.default.url(forUbiquityContainerIdentifier:nil)?.appendingPathComponent("Documents").appendingPathComponent(fileName, isDirectory: false)
-    
-    if (iCloudDocumentsURL?.absoluteString.isEmpty == false){
-        print("iCloud Drive loaded file from \(iCloudDocumentsURL?.path) and copying to \(localFile)")
+    if (loadingiCloud == false){
+        loadingiCloud = true;
+        let iCloudDocumentsURL = FileManager.default.url(forUbiquityContainerIdentifier:nil)?.appendingPathComponent("Documents").appendingPathComponent(fileName, isDirectory: false)
         
-        do {
-            try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
-            let localModify : Date = try FileManager.default.attributesOfItem(atPath: localFile.path)[FileAttributeKey.modificationDate] as! Date;
+        if (iCloudDocumentsURL?.absoluteString.isEmpty == false){
+            print("iCloud Drive loaded file from \(iCloudDocumentsURL?.path) and copying to \(localFile)")
             
-            let cloudModify :Date  = try FileManager.default.attributesOfItem(atPath: iCloudDocumentsURL!.path)[FileAttributeKey.modificationDate] as! Date;
-            
-            //print ("iCloud Drive localModify is \(localModify) and cloudModify is \(cloudModify)")
-            //if (cloudModify > localModify){
-            
-            if (FileManager.default.fileExists(atPath: (localFile.path)) == true){
-                try FileManager.default.removeItem(atPath: localFile.path)
+            do {
+                try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
+                let localModify : Date = try FileManager.default.attributesOfItem(atPath: localFile.path)[FileAttributeKey.modificationDate] as! Date;
+                
+                let cloudModify :Date  = try FileManager.default.attributesOfItem(atPath: iCloudDocumentsURL!.path)[FileAttributeKey.modificationDate] as! Date;
+                
+                //print ("iCloud Drive localModify is \(localModify) and cloudModify is \(cloudModify)")
+                //if (cloudModify > localModify){
+                
+                if (FileManager.default.fileExists(atPath: (localFile.path)) == true){
+                    try FileManager.default.removeItem(atPath: localFile.path)
+                }
+                print("iCloud Drive copied file \(iCloudDocumentsURL?.path) to \(localFile.path)")
+                try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
+                try FileManager.default.copyItem(atPath: (iCloudDocumentsURL?.path)!, toPath: localFile.path)
+                print("iCloud Drive wrote file \(localFile.path)")
+                
+            } catch {
+                print("iCloud Drive error on load of single file of \(iCloudDocumentsURL?.path) could not be copied to \(localFile.path) " + error.localizedDescription);
             }
-            print("iCloud Drive copied file \(iCloudDocumentsURL?.path) to \(localFile.path)")
-            try FileManager.default.startDownloadingUbiquitousItem(at: iCloudDocumentsURL!)
-            try FileManager.default.copyItem(atPath: (iCloudDocumentsURL?.path)!, toPath: localFile.path)
-            print("iCloud Drive wrote file \(localFile.path)")
-            
-        } catch {
-            print("iCloud Drive error on load of single file of \(iCloudDocumentsURL?.path) could not be copied to \(localFile.path) " + error.localizedDescription);
+        } else {
+            print("iCloud Drive file \(localFile.path) is not available in the cloud")
         }
-    } else {
-        print("iCloud Drive file \(localFile.path) is not available in the cloud")
+        loadingiCloud = false;
     }
 }
